move Not into _not.rs
