+++ api/src/term/matcher.rs
@@ -21,14 +21,16 @@ mod _datatype_matcher;
 mod _graph_name_matcher;
 mod _language_tag_matcher;
 mod _matcher_ref;
+mod _not;
 mod _term_matcher_gn;
 mod _trait;
 
-pub use _any::*;
+pub use _any::Any;
 pub use _datatype_matcher::*;
 pub use _graph_name_matcher::*;
 pub use _language_tag_matcher::*;
 pub use _matcher_ref::*;
+pub use _not::Not;
 pub use _term_matcher_gn::*;
 pub use _trait::*;
